requirements.txt
requests>=2.31
python-dotenv>=1.0
SQLAlchemy>=2.0
RapidFuzz>=3.0
